Changing Variant to void *
This save a lot of memory in expression trees, while keeping the ability to perform a similar union cast (with pointers).

+++ engine/core/primitive.hpp
@@ -11,7 +11,7 @@ namespace zhetapi {
 #define MAX_PIDS	3
 
 // ID type alias
-using TypeId = size_t;
+using TypeId = uint32_t;
 
 // Enode type IDs (at most 16) TODO: put inside primitive
 enum PrimIds : TypeId {
